+++ bundlers/babel/BabelBundler.ts
@@ -14,8 +14,14 @@ module BabelBundler {
 
     /** Create a browserify transform which compiles source files using babelify
      */
-    export function createTransformer(babelify: babelify.BabelifyConstructor, filePattern?: { test(str: string): boolean; } | RegExp,
-            babelCompilerOpts?: any, transformOpts?: BrowserifyHelper.BrowserifyTransform["options"], verbose?: boolean) {
+    export function createTransformer(
+        babelify: babelify.BabelifyConstructor,
+        filePattern?: { test(str: string): boolean; } | RegExp,
+        babelCompilerOpts?: babelify.BabelifyOptions,
+        transformOpts?: BrowserifyHelper.BrowserifyTransform["options"],
+        verbose?: boolean | ((...args: any[]) => void)
+    ) {
+        var log = (typeof verbose === "function" ? verbose : verbose == true ? console.log : null);
 
         var res: BrowserifyHelper.BrowserifyTransform = {
             transform: function babelifyTransform(tr, opts) {
@@ -23,8 +29,8 @@ module BabelBundler {
                     return new stream.PassThrough();
                 }
 
-                if (verbose) {
-                    console.log("babelify: '" + PathUtil.toShortFileName(tr) + "'");
+                if (log != null) {
+                    log("babelify: ", PathUtil.toShortFileName(tr));
                 }
 
                 var strm = babelify(tr, BrowserifyHelper.combineOpts(opts, babelCompilerOpts));