+++ test/test-proj/App.js
@@ -1,7 +1,11 @@
 "use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
+var DataSource = require("./DataSource");
 var HelperUtil = require("./HelperUtil");
 function main() {
     console.log(HelperUtil.App.name + "@" + HelperUtil.App.version);
+    for (var collName in DataSource.collectionCache) {
+        console.log(collName + ": " + DataSource.collectionCache[collName]);
+    }
 }
 main();
+module.exports = main;